Add more vanilla style selection boxes
